resume world 1x1

+++ continuous/thailand/other_variable/main.py
@@ -23,6 +23,7 @@ def main():
     else:
         print(f"train_flag is {train_flag}: not saved")
 
+    px.validation()
     px.show(val_index=px.val_index)
 
 class Pixel():
@@ -74,34 +75,42 @@ class Pixel():
             data = pickle.load(f)
         x_val, y_val = data['x_val'], data['y_val']
 
-        pred_lst = []
-        rmse = []
-        corr = []
-        rr = []
-        for i in range(self.grid_num):
-            y_val_px = y_val[:, i]
-            model = build_model((self.lat, self.lon, self.var_num))
-            model.compile(optimizer=self.optimizer, loss=self.loss, metrics=[self.metrics])
-            weights_path = f"{self.weights_dir}/epoch{self.epochs}_batch{self.batch_size}_{i}.h5"
-            model.load_weights(weights_path)
+        if os.path.exists(self.result_path) is False:
+            pred_lst = []
+            rmse = []
+            corr = []
+            rr = []
+            for i in range(self.grid_num):
+                y_val_px = y_val[:, i]
+                model = build_model((self.lat, self.lon, self.var_num))
+                model.compile(optimizer=self.optimizer, loss=self.loss, metrics=[self.metrics])
+                weights_path = f"{self.weights_dir}/epoch{self.epochs}_batch{self.batch_size}_{i}.h5"
+                model.load_weights(weights_path)
 
-            pred = model.predict(x_val) # (400, 1000)
-            pred_lst.append(pred)
+                pred = model.predict(x_val) # (400, 1000)
+                pred_lst.append(pred)
 
-            result = model.evaluate(x_val, y_val_px)
-            rmse.append(round(result[1], 2))
+                result = model.evaluate(x_val, y_val_px)
+                rmse.append(round(result[1], 2))
 
-            pred = model.predict(x_val)
-            corr_i = np.corrcoef(pred[:,0], y_val_px)
-            corr.append(np.round(corr_i[0,1], 2))
+                pred = model.predict(x_val)
+                corr_i = np.corrcoef(pred[:,0], y_val_px)
+                corr.append(np.round(corr_i[0,1], 2))
 
-            rr_i = corr_i**2
-            rr.append(np.round(rr_i, 2))
+                rr_i = corr_i**2
+                rr.append(np.round(rr_i, 2))
 
-            print(f"Correlation Coefficient of pixel{i}: {np.round(corr_i[0,1], 2)}")
-
-        pred_arr = np.array(pred_lst)
-        np.save(self.result_path, pred_arr)
+                print(f"Correlation Coefficient of pixel{i}: {np.round(corr_i[0,1], 2)}")
+                os.makedirs(self.result_dir, )
+                pred_arr = np.array(pred_lst)
+                np.save(self.result_path, pred_arr)
+        else:
+            corr = []
+            pred_arr = np.squeeze(np.load(self.result_path))
+            for i in range(self.grid_num):
+                y_val_px = y_val[:, i]
+                corr_i = np.corrcoef(pred_arr[i, :], y_val_px)
+                corr.append(np.round(corr_i[0,1], 2))
 
         corr = np.array(corr)
         corr = corr.reshape(self.lat_grid, self.lon_grid)
